Refactoring Device type out of the module
