+++ android/src/main/java/com/learnium/RNDeviceInfo/RNDeviceModule.java
@@ -3,7 +3,6 @@ package com.learnium.RNDeviceInfo;
 import android.Manifest;
 import android.annotation.SuppressLint;
 import android.app.KeyguardManager;
-import android.app.UiModeManager;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -12,7 +11,6 @@ import android.content.IntentFilter;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.pm.FeatureInfo;
-import android.content.res.Configuration;
 import android.location.LocationManager;
 import android.net.wifi.WifiManager;
 import android.net.wifi.WifiInfo;
@@ -22,12 +20,10 @@ import android.os.PowerManager;
 import android.os.StatFs;
 import android.os.BatteryManager;
 import android.provider.Settings;
-import android.view.WindowManager;
 import android.webkit.WebSettings;
 import android.telephony.TelephonyManager;
 import android.text.TextUtils;
 import android.app.ActivityManager;
-import android.util.DisplayMetrics;
 import android.hardware.Camera;
 import android.hardware.camera2.CameraManager;
 
@@ -65,6 +61,7 @@ import static android.provider.Settings.Secure.getString;
 @ReactModule(name = RNDeviceModule.NAME)
 public class RNDeviceModule extends ReactContextBaseJavaModule {
   public static final String NAME = "RNDeviceInfo";
+  private final DeviceTypeResolver deviceTypeResolver;
   private BroadcastReceiver receiver;
 
   private double mLastBatteryLevel = -1;
@@ -76,6 +73,7 @@ public class RNDeviceModule extends ReactContextBaseJavaModule {
 
   public RNDeviceModule(ReactApplicationContext reactContext) {
     super(reactContext);
+    this.deviceTypeResolver = new DeviceTypeResolver(reactContext);
   }
 
   @Override
@@ -160,11 +158,11 @@ public class RNDeviceModule extends ReactContextBaseJavaModule {
     constants.put("systemVersion", Build.VERSION.RELEASE);
     constants.put("appVersion", appVersion);
     constants.put("buildNumber", buildNumber);
-    constants.put("isTablet", getDeviceType() == DeviceType.TABLET);
+    constants.put("isTablet", deviceTypeResolver.isTablet());
     constants.put("appName", appName);
     constants.put("brand", Build.BRAND);
     constants.put("model", Build.MODEL);
-    constants.put("deviceType", getDeviceType().getValue());
+    constants.put("deviceType", deviceTypeResolver.getDeviceType().getValue());
 
     return constants;
   }
@@ -200,72 +198,6 @@ public class RNDeviceModule extends ReactContextBaseJavaModule {
             || Build.PRODUCT.toLowerCase().contains("nox")
             || Build.SERIAL.toLowerCase().contains("nox")
             || (Build.BRAND.startsWith("generic") && Build.DEVICE.startsWith("generic"));
-  }
-
-  private DeviceType getDeviceType() {
-    // Detect TVs via ui mode (Android TVs) or system features (Fire TV).
-    if (getReactApplicationContext().getPackageManager().hasSystemFeature("amazon.hardware.fire_tv")) {
-      return DeviceType.TV;
-    }
-
-    UiModeManager uiManager = (UiModeManager) getReactApplicationContext().getSystemService(Context.UI_MODE_SERVICE);
-    if (uiManager != null && uiManager.getCurrentModeType() == Configuration.UI_MODE_TYPE_TELEVISION) {
-      return DeviceType.TV;
-    }
-
-    DeviceType deviceTypeFromConfig = getDeviceTypeFromResourceConfiguration();
-
-    if (deviceTypeFromConfig != null && deviceTypeFromConfig != DeviceType.UNKNOWN) {
-      return deviceTypeFromConfig;
-    }
-
-    return  getDeviceTypeFromPhysicalSize();
-  }
-
-  // Use `smallestScreenWidthDp` to determine the screen size
-  // https://android-developers.googleblog.com/2011/07/new-tools-for-managing-screen-sizes.html
-  private DeviceType getDeviceTypeFromResourceConfiguration() {
-    int smallestScreenWidthDp = getReactApplicationContext().getResources().getConfiguration().smallestScreenWidthDp;
-
-    if (smallestScreenWidthDp == Configuration.SMALLEST_SCREEN_WIDTH_DP_UNDEFINED) {
-      return DeviceType.UNKNOWN;
-    }
-
-    return smallestScreenWidthDp >= 600 ? DeviceType.TABLET : DeviceType.HANDSET;
-  }
-
-  private DeviceType getDeviceTypeFromPhysicalSize() {
-    // Find the current window manager, if none is found we can't measure the device physical size.
-    WindowManager windowManager = (WindowManager) getReactApplicationContext().getSystemService(Context.WINDOW_SERVICE);
-
-    if (windowManager == null) {
-      return DeviceType.UNKNOWN;
-    }
-
-    // Get display metrics to see if we can differentiate handsets and tablets.
-    // NOTE: for API level 16 the metrics will exclude window decor.
-    DisplayMetrics metrics = new DisplayMetrics();
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
-      windowManager.getDefaultDisplay().getRealMetrics(metrics);
-    } else {
-      windowManager.getDefaultDisplay().getMetrics(metrics);
-    }
-
-    // Calculate physical size.
-    double widthInches = metrics.widthPixels / (double) metrics.xdpi;
-    double heightInches = metrics.heightPixels / (double) metrics.ydpi;
-    double diagonalSizeInches = Math.sqrt(Math.pow(widthInches, 2) + Math.pow(heightInches, 2));
-
-    if (diagonalSizeInches >= 3.0 && diagonalSizeInches <= 6.9) {
-      // Devices in a sane range for phones are considered to be Handsets.
-      return DeviceType.HANDSET;
-    } else if (diagonalSizeInches > 6.9 && diagonalSizeInches <= 18.0) {
-      // Devices larger than handset and in a sane range for tablets are tablets.
-      return DeviceType.TABLET;
-    } else {
-      // Otherwise, we don't know what device type we're on/
-      return DeviceType.UNKNOWN;
-    }
   }
 
   @ReactMethod(isBlockingSynchronousMethod = true)
